refactor(alloy-celo-evm): share CIP-64 receipt construction between alloy & reth builders
CeloAlloyReceiptBuilder and CeloRethReceiptBuilder previously duplicated
the CIP-64 receipt branch: pop the pending entry, assert the
data-present-on-success invariant, merge pre/post logs into the main
execution logs, and thread base_fee_in_erc20 into a CeloCip64Receipt.
The only real divergence is the outer wrapping — bloomed
CeloReceiptEnvelope::Cip64 for canonical/alloy types vs. bloomless
CeloReceipt::Cip64 for reth's on-disk storage.

Move the shared inner construction onto Cip64Storage as
`build_cip64_receipt`. Both builders' CIP-64 arms become one-liners that
wrap the returned CeloCip64Receipt in their envelope variant (with or
without `.with_bloom()`). The previously duplicated `merge_logs` helper
is inlined at its sole remaining caller.
